restructured project, should have no problem running
