+++ dating/__init__.py
@@ -1,0 +1,11 @@
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+
+app = Flask(__name__)
+
+app.config['SECRET_KEY'] = '7ba88159506e7b84bff4420080f75a92'
+app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///site.db'
+
+db = SQLAlchemy(app)
+
+from dating import routes
