+++ dating/templates/index.html
@@ -1,5 +1,9 @@
 { % extends "layout.html" %}
 
+{% block title %}
+    <title> HomePage </title>
+{% endblock title %}
+
 { % block content %}
-  
+    
 { % endblock content %}
