+++ run.py
@@ -1,0 +1,5 @@
+from dating import app
+
+
+if __name__ == '__main__':
+    app.run(debug=True)
